Refactor WorkList and WorkListId.
