+++ src/main/java/seedu/address/model/Model.java
@@ -7,7 +7,6 @@ import javafx.beans.property.ReadOnlyProperty;
 import javafx.collections.ObservableList;
 import seedu.address.commons.core.GuiSettings;
 import seedu.address.model.equipment.Equipment;
-import seedu.address.model.equipment.WorkList;
 import seedu.address.model.tag.Tag;
 
 /**
