+++ src/test/java/seedu/address/logic/commands/AddCommandTest.java
@@ -24,7 +24,7 @@ import seedu.address.model.Model;
 import seedu.address.model.ReadOnlyEquipmentManager;
 import seedu.address.model.ReadOnlyUserPrefs;
 import seedu.address.model.equipment.Equipment;
-import seedu.address.model.equipment.WorkList;
+import seedu.address.model.WorkList;
 import seedu.address.model.tag.Tag;
 import seedu.address.testutil.EquipmentBuilder;
