+++ src/test/java/seedu/address/model/equipment/UniqueWorkListListTest.java
@@ -21,6 +21,7 @@ import org.junit.Rule;
 import org.junit.Test;
 import org.junit.rules.ExpectedException;
 
+import seedu.address.model.WorkList;
 import seedu.address.model.equipment.exceptions.DuplicateEquipmentException;
 import seedu.address.model.equipment.exceptions.EquipmentNotFoundException;
 //import seedu.address.testutil.EquipmentBuilder;
